perf: improve codec handling in load_filepaths_and_text function in `infer.lib.train.utils`
this change incorporates usage of `codecs` module to dynamically guess the encoding of the file and convert it to pure utf-8. results of this convertion should be pure and work as intended. the idea is to not to open the file AGAIN if the UnicodeDecodeError occurs. The only possible downside is that if the encoding can not be translated to utf-8 directly (rare case) - the file will not be properly opened at all.